requirements.txt
flask
flask-cors
python-dotenv
google-genai
numpy
